feat: move Debug tab to TopBar button with full sidebar
Removes the Debug tab from ChatPane's tab strip and adds a Debug button
(⌥) to the TopBar that toggles a full-height sidebar overlay. The sidebar
shows foreman conversation context with the same refresh/clear controls,
closeable via X button or clicking the overlay backdrop.

Closes #131
